Refactor sendGrantEmail into a more reuseable function
