+++ src/data/emailTemplates/review.handlebars
@@ -17,7 +17,6 @@
     <p>askAmount: {{askAmount}}</p>
     <p>totalCost: {{totalCost}}</p>
     <p>Please review it as quickly as possible.</p>
-    <p><a href="https://communityfund.co/email-reviewer/{{id}}/accept">Accept</a></p>
-    <p><a href="https://communityfund.co/email-reviewer/{{id}}/reject">Reject</a></p>
+    <p><a href="{{reviewUrl}}">Review</a></p>
   </body>
 </html>
