+++ src/data/mutations/createSubmission.js
@@ -4,11 +4,39 @@ import {
 } from 'graphql';
 import moment from 'moment';
 import SubmissionType from '../types/SubmissionType';
-import { Submission, User } from '../models';
+import { Submission, User, Vote } from '../models';
 import { logError } from '../../logger';
 import { isDev, googleRecaptchaSecret } from '../../config';
 import fetch from '../../core/fetch';
-import { sendReviewEmail } from '../../core/email';
+import reviewTemplate from '../emailTemplates/review.handlebars';
+import sendEmail from '../../core/email';
+
+export const sendEmailToReviewer = async (reviewerEmail, submissionId) => {
+  console.log('Sending grant review email', reviewerEmail, submissionId);
+
+  const submission = await Submission.findOne({
+    id: submissionId,
+  });
+
+  if (!submission) {
+    throw Error('Submission not found');
+  }
+
+  // TODO: We need to read this dynamically
+  const hostName = 'http://communityfund.co';
+
+  const html = reviewTemplate({
+    id: submission.id,
+    name: submission.name,
+    phone: submission.phone,
+    email: submission.email,
+    askAmount: submission.askAmount,
+    totalCost: submission.totalCost,
+    reviewUrl: `${hostName}/submission/${submission.id}`,
+  });
+
+  sendEmail(reviewerEmail, 'New grant application', html);
+};
 
 const createSubmission = {
   type: SubmissionType,
@@ -23,17 +51,20 @@ const createSubmission = {
     totalCost: { type: new NonNull(StringType) },
     googleToken: { type: new NonNull(StringType) },
   },
-  resolve: async ({ req }, {
-    name,
-    email,
-    phone,
-    date,
-    summary,
-    description,
-    askAmount,
-    totalCost,
-    googleToken,
-  }) => {
+  resolve: async (
+    { req },
+    {
+      name,
+      email,
+      phone,
+      date,
+      summary,
+      description,
+      askAmount,
+      totalCost,
+      googleToken,
+    }
+  ) => {
     // Only validate with recaptcha if we are on prod. To Pass the authentication
     // the googleRecaptchaSecret needs to be passed into the process.
     if (!isDev) {
@@ -79,12 +110,16 @@ const createSubmission = {
       throw e;
     }
 
+    // TODO: Send email to applicant
+
     const reviewers = await User.findAll({
       isReviewer: true,
     });
 
     // Send review email to all reviewers
-    await Promise.all(reviewers.map(user => sendReviewEmail(user.email, submission.id)));
+    await Promise.all(
+      reviewers.map(user => sendEmailToReviewer(user.email, submission.id))
+    );
 
     return submission;
   },
